Split the command line functions into several themed modules and renamed "sgr file" and "sgr rerun" to "sgr build" and "sgr rebuild".
