+++ splitgraph/commandline/console.py
